+++ src/algorithms/local_search/local_search.cpp
@@ -92,11 +92,6 @@ LocalSearch<Route,
 
   // Setup solution state.
   _sol_state.setup(_sol);
-
-#ifdef LOG_LS_OPERATORS
-  tried_moves.fill(0);
-  applied_moves.fill(0);
-#endif
 }
 
 template <class Route>
@@ -142,30 +137,31 @@ template <class Route,
           class RouteSplit,
           class PriorityReplace,
           class TSPFix>
-void LocalSearch<Route,
-                 UnassignedExchange,
-                 CrossExchange,
-                 MixedExchange,
-                 TwoOpt,
-                 ReverseTwoOpt,
-                 Relocate,
-                 OrOpt,
-                 IntraExchange,
-                 IntraCrossExchange,
-                 IntraMixedExchange,
-                 IntraRelocate,
-                 IntraOrOpt,
-                 IntraTwoOpt,
-                 PDShift,
-                 RouteExchange,
-                 SwapStar,
-                 RouteSplit,
-                 PriorityReplace,
-                 TSPFix>::try_job_additions(const std::vector<Index>& routes,
-                                            double regret_coeff
+std::unordered_set<Index>
+LocalSearch<Route,
+            UnassignedExchange,
+            CrossExchange,
+            MixedExchange,
+            TwoOpt,
+            ReverseTwoOpt,
+            Relocate,
+            OrOpt,
+            IntraExchange,
+            IntraCrossExchange,
+            IntraMixedExchange,
+            IntraRelocate,
+            IntraOrOpt,
+            IntraTwoOpt,
+            PDShift,
+            RouteExchange,
+            SwapStar,
+            RouteSplit,
+            PriorityReplace,
+            TSPFix>::try_job_additions(const std::vector<Index>& routes,
+                                       double regret_coeff
 #ifdef LOG_LS
-                                            ,
-                                            bool log_addition_step
+                                       ,
+                                       bool log_addition_step
 #endif
 ) {
   bool job_added;
@@ -336,9 +332,20 @@ void LocalSearch<Route,
     }
   } while (job_added);
 
+  // Update stored data for consistency (except update_route_eval and
+  // set_insertion_ranks done along the way).
   for (const auto v : modified_vehicles) {
     _sol_state.update_route_bbox(_sol[v].route, v);
+    _sol_state.update_costs(_sol[v].route, v);
+    _sol_state.update_skills(_sol[v].route, v);
+    _sol_state.update_priorities(_sol[v].route, v);
+    _sol_state.set_node_gains(_sol[v].route, v);
+    _sol_state.set_edge_gains(_sol[v].route, v);
+    _sol_state.set_pd_matching_ranks(_sol[v].route, v);
+    _sol_state.set_pd_gains(_sol[v].route, v);
   }
+
+  return modified_vehicles;
 }
 
 template <class Route,
@@ -500,9 +507,7 @@ void LocalSearch<Route,
                   // Same move as with t_rank == s_rank.
                   continue;
                 }
-#ifdef LOG_LS_OPERATORS
-                ++tried_moves[OperatorName::UnassignedExchange];
-#endif
+
                 UnassignedExchange r(_input,
                                      _sol_state,
                                      _sol_state.unassigned,
@@ -603,9 +608,6 @@ void LocalSearch<Route,
 
           if (best_current_priority > 0 &&
               best_priorities[source] <= best_current_priority) {
-#ifdef LOG_LS_OPERATORS
-            ++tried_moves[OperatorName::PriorityReplace];
-#endif
             PriorityReplace r(_input,
                               _sol_state,
                               _sol_state.unassigned,
@@ -752,9 +754,6 @@ void LocalSearch<Route,
             continue;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::CrossExchange];
-#endif
           CrossExchange r(_input,
                           _sol_state,
                           _sol[source],
@@ -869,9 +868,6 @@ void LocalSearch<Route,
               continue;
             }
 
-#ifdef LOG_LS_OPERATORS
-            ++tried_moves[OperatorName::MixedExchange];
-#endif
             MixedExchange r(_input,
                             _sol_state,
                             _sol[source],
@@ -983,9 +979,6 @@ void LocalSearch<Route,
             continue;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::TwoOpt];
-#endif
           TwoOpt r(_input,
                    _sol_state,
                    _sol[source],
@@ -1086,9 +1079,6 @@ void LocalSearch<Route,
             continue;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::ReverseTwoOpt];
-#endif
           ReverseTwoOpt r(_input,
                           _sol_state,
                           _sol[source],
@@ -1152,9 +1142,6 @@ void LocalSearch<Route,
                  _sol_state.insertion_ranks_begin[target][s_job_rank];
                t_rank < _sol_state.insertion_ranks_end[target][s_job_rank];
                ++t_rank) {
-#ifdef LOG_LS_OPERATORS
-            ++tried_moves[OperatorName::Relocate];
-#endif
             Relocate r(_input,
                        _sol_state,
                        _sol[source],
@@ -1228,9 +1215,6 @@ void LocalSearch<Route,
                      _sol_state.insertion_ranks_end[target][s_next_job_rank]);
           for (unsigned t_rank = insertion_start; t_rank < insertion_end;
                ++t_rank) {
-#ifdef LOG_LS_OPERATORS
-            ++tried_moves[OperatorName::OrOpt];
-#endif
             OrOpt r(_input,
                     _sol_state,
                     _sol[source],
@@ -1259,9 +1243,6 @@ void LocalSearch<Route,
           continue;
         }
 
-#ifdef LOG_LS_OPERATORS
-        ++tried_moves[OperatorName::TSPFix];
-#endif
         TSPFix op(_input, _sol_state, _sol[source], source);
 
         if (best_gains[source][target] < op.gain() && op.is_valid()) {
@@ -1306,9 +1287,6 @@ void LocalSearch<Route,
             continue;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::IntraExchange];
-#endif
           IntraExchange r(_input,
                           _sol_state,
                           _sol[source],
@@ -1376,9 +1354,6 @@ void LocalSearch<Route,
             continue;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::IntraCrossExchange];
-#endif
           IntraCrossExchange r(_input,
                                _sol_state,
                                _sol[source],
@@ -1448,9 +1423,6 @@ void LocalSearch<Route,
             continue;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::IntraMixedExchange];
-#endif
           IntraMixedExchange r(_input,
                                _sol_state,
                                _sol[source],
@@ -1511,9 +1483,6 @@ void LocalSearch<Route,
             break;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::IntraRelocate];
-#endif
           IntraRelocate r(_input,
                           _sol_state,
                           _sol[source],
@@ -1586,9 +1555,6 @@ void LocalSearch<Route,
             break;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::IntraOrOpt];
-#endif
           IntraOrOpt r(_input,
                        _sol_state,
                        _sol[source],
@@ -1621,9 +1587,6 @@ void LocalSearch<Route,
                                    static_cast<Index>(end_s - 1));
 
         for (unsigned t_rank = s_rank + 2; t_rank < end_t_rank; ++t_rank) {
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::IntraTwoOpt];
-#endif
           IntraTwoOpt r(_input,
                         _sol_state,
                         _sol[source],
@@ -1688,9 +1651,6 @@ void LocalSearch<Route,
             continue;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::PDShift];
-#endif
           PDShift pdr(_input,
                       _sol_state,
                       _sol[source],
@@ -1743,9 +1703,6 @@ void LocalSearch<Route,
           continue;
         }
 
-#ifdef LOG_LS_OPERATORS
-        ++tried_moves[OperatorName::RouteExchange];
-#endif
         RouteExchange re(_input,
                          _sol_state,
                          _sol[source],
@@ -1775,9 +1732,6 @@ void LocalSearch<Route,
           continue;
         }
 
-#ifdef LOG_LS_OPERATORS
-        ++tried_moves[OperatorName::SwapStar];
-#endif
         SwapStar r(_input,
                    _sol_state,
                    _sol[source],
@@ -1812,9 +1766,6 @@ void LocalSearch<Route,
             continue;
           }
 
-#ifdef LOG_LS_OPERATORS
-          ++tried_moves[OperatorName::RouteSplit];
-#endif
           // RouteSplit stores a const& to empty_route_ranks, which
           // will be invalid in the unique_ptr created below after
           // empty_route_ranks goes out of scope. This is fine because
@@ -1876,10 +1827,6 @@ void LocalSearch<Route,
       auto update_candidates =
         best_ops[best_source][best_target]->update_candidates();
 
-#ifdef LOG_LS_OPERATORS
-      ++applied_moves.at(best_ops[best_source][best_target]->get_name());
-#endif
-
 #ifdef LOG_LS
       steps.emplace_back(utils::now(),
                          log::EVENT::OPERATOR,
@@ -1902,6 +1849,14 @@ void LocalSearch<Route,
       for (auto v_rank : update_candidates) {
         _sol_state.update_route_eval(_sol[v_rank].route, v_rank);
         _sol_state.update_route_bbox(_sol[v_rank].route, v_rank);
+        _sol_state.update_costs(_sol[v_rank].route, v_rank);
+        _sol_state.update_skills(_sol[v_rank].route, v_rank);
+        _sol_state.update_priorities(_sol[v_rank].route, v_rank);
+        _sol_state.set_insertion_ranks(_sol[v_rank], v_rank);
+        _sol_state.set_node_gains(_sol[v_rank].route, v_rank);
+        _sol_state.set_edge_gains(_sol[v_rank].route, v_rank);
+        _sol_state.set_pd_matching_ranks(_sol[v_rank].route, v_rank);
+        _sol_state.set_pd_gains(_sol[v_rank].route, v_rank);
 
         assert(_sol[v_rank].size() <= _input.vehicles[v_rank].max_tasks);
         assert(_input.vehicles[v_rank].ok_for_range_bounds(
@@ -1919,29 +1874,24 @@ void LocalSearch<Route,
       assert(new_eval + best_gain == previous_eval);
 #endif
 
-      for (auto v_rank : update_candidates) {
-        // Only this update (and update_route_eval done above) are
-        // actually required for consistency inside try_job_additions.
-        _sol_state.set_insertion_ranks(_sol[v_rank], v_rank);
-      }
-
-      try_job_additions(best_ops[best_source][best_target]
-                          ->addition_candidates(),
-                        0
+      auto modified_vehicles =
+        try_job_additions(best_ops[best_source][best_target]
+                            ->addition_candidates(),
+                          0
 #ifdef LOG_LS
-                        ,
-                        true
+                          ,
+                          true
 #endif
-      );
+        );
 
-      for (auto v_rank : update_candidates) {
-        _sol_state.update_costs(_sol[v_rank].route, v_rank);
-        _sol_state.update_skills(_sol[v_rank].route, v_rank);
-        _sol_state.update_priorities(_sol[v_rank].route, v_rank);
-        _sol_state.set_node_gains(_sol[v_rank].route, v_rank);
-        _sol_state.set_edge_gains(_sol[v_rank].route, v_rank);
-        _sol_state.set_pd_matching_ranks(_sol[v_rank].route, v_rank);
-        _sol_state.set_pd_gains(_sol[v_rank].route, v_rank);
+      // Extend update_candidates in case a vehicle was not modified
+      // by the operator itself but afterward by
+      // try_job_additions. Can happen e.g. with UnassignedExchange.
+      for (const auto v : update_candidates) {
+        modified_vehicles.erase(v);
+      }
+      for (const auto v : modified_vehicles) {
+        update_candidates.push_back(v);
       }
 
       // Set gains to zero for what needs to be recomputed in the next
@@ -2118,26 +2068,19 @@ void LocalSearch<Route,
                          utils::format_solution(_input, _sol));
 #endif
 
-      // Update insertion ranks ranges.
+      // Update stored data that has not been maintained while
+      // removing.
       for (std::size_t v = 0; v < _sol.size(); ++v) {
+        _sol_state.update_costs(_sol[v].route, v);
+        _sol_state.update_skills(_sol[v].route, v);
+        _sol_state.update_priorities(_sol[v].route, v);
         _sol_state.set_insertion_ranks(_sol[v], v);
+        _sol_state.set_edge_gains(_sol[v].route, v);
       }
 
       // Refill jobs.
       constexpr double refill_regret = 1.5;
       try_job_additions(_all_routes, refill_regret);
-
-      // Update everything except what has already been updated in
-      // try_job_additions.
-      for (std::size_t v = 0; v < _sol.size(); ++v) {
-        _sol_state.update_costs(_sol[v].route, v);
-        _sol_state.update_skills(_sol[v].route, v);
-        _sol_state.update_priorities(_sol[v].route, v);
-        _sol_state.set_node_gains(_sol[v].route, v);
-        _sol_state.set_edge_gains(_sol[v].route, v);
-        _sol_state.set_pd_matching_ranks(_sol[v].route, v);
-        _sol_state.set_pd_gains(_sol[v].route, v);
-      }
 
 #ifdef LOG_LS
       steps.emplace_back(utils::now(),
@@ -2149,57 +2092,6 @@ void LocalSearch<Route,
     }
   }
 }
-
-#ifdef LOG_LS_OPERATORS
-template <class Route,
-          class UnassignedExchange,
-          class CrossExchange,
-          class MixedExchange,
-          class TwoOpt,
-          class ReverseTwoOpt,
-          class Relocate,
-          class OrOpt,
-          class IntraExchange,
-          class IntraCrossExchange,
-          class IntraMixedExchange,
-          class IntraRelocate,
-          class IntraOrOpt,
-          class IntraTwoOpt,
-          class PDShift,
-          class RouteExchange,
-          class SwapStar,
-          class RouteSplit,
-          class PriorityReplace,
-          class TSPFix>
-std::array<OperatorStats, OperatorName::MAX>
-LocalSearch<Route,
-            UnassignedExchange,
-            CrossExchange,
-            MixedExchange,
-            TwoOpt,
-            ReverseTwoOpt,
-            Relocate,
-            OrOpt,
-            IntraExchange,
-            IntraCrossExchange,
-            IntraMixedExchange,
-            IntraRelocate,
-            IntraOrOpt,
-            IntraTwoOpt,
-            PDShift,
-            RouteExchange,
-            SwapStar,
-            RouteSplit,
-            PriorityReplace,
-            TSPFix>::get_stats() const {
-  std::array<OperatorStats, OperatorName::MAX> stats;
-  for (auto op = 0; op < OperatorName::MAX; ++op) {
-    stats[op] = OperatorStats(tried_moves.at(op), applied_moves.at(op));
-  }
-
-  return stats;
-}
-#endif
 
 template <class Route,
           class UnassignedExchange,